requirements.txt
# API
fastapi==0.111.0
uvicorn==0.30.1

# ML
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4

# MLOps
mlflow==2.13.0
evidently==0.4.30

# Observabilidad
prometheus-fastapi-instrumentator==7.0.0

# Utilidades
python-dotenv==1.0.1
jinja2==3.1.4